Removed influxdb 8 support. Added InfluxDB 9 support.
