+++ scripts/influxdb_run.sh
@@ -1,1 +1,1 @@
-docker run -d -p 8083:8083 -p 8086:8086 --expose 8090 --expose 8099 tutum/influxdb
+docker run -d -p 8083:8083 -p 8086:8086 --expose 8083 --expose 8086 tutum/influxdb:0.9
