+++ src/tests/time_series_export/slack_ts_export_test.cpp
@@ -1,9 +1,9 @@
 #include "observers/slack_resource.hpp"
 
 #include "tests/common/sources/json_source.hpp"
-#include "tests/common/sinks/mock_sink.hpp"
+#include "tests/common/mocks/mock_sink.hpp"
 
-#include "time_series_export/backend/influx_db8.hpp"
+#include "time_series_export/backend/influx_db9.hpp"
 #include "time_series_export/slack_ts_export.hpp"
 
 namespace mesos {
@@ -12,8 +12,8 @@ namespace tests {
 
 TEST(SlackResourceTimeSeriesExportTest, BasicTest) {
   SlackResourceObserver observer;
-  InfluxDb8Backend backend("localhost", "8086", "serenity", "root", "root");
-  SlackTimeSeriesExporter slackExporter("tagged-test", backend);
+  InfluxDb9Backend backend("localhost", "8086", "serenity", "root", "root");
+  SlackTimeSeriesExporter slackExporter("tagged-test", &backend);
   JsonSource jsonSource;
   MockSink<Resources> mockSink;
